Added recent items filtered by library and cleaned some stuff up
